# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5Swagger",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local"
        },
        {
            "url": "http://localhost/api/auth",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/transaction": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получить все транзакции",
                "description": "Получить транзакции",
                "operationId": "0d0da50770e8f47be81154ef4e663a65",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transaction/{id}": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получить транзакцию",
                "description": "Получает транзакцию по идентификатору и возвращает её",
                "operationId": "aa7f4f0f95b8cf5c3b2131eea26bf731",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1bbf9e65-a2ec-46e8-b155-5dac60831817"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transaction/create": {
            "post": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Создать транзакцию",
                "description": "Создает новую транзакцию и возвращает её",
                "operationId": "1ca56c697366694665bd78a005137594",
                "parameters": [
                    {
                        "name": "arendator_id",
                        "in": "path",
                        "description": "Идентификатор арендатора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "8f9a9586-4879-39d7-9040-423e5297cbc8"
                        }
                    },
                    {
                        "name": "bill_id",
                        "in": "path",
                        "description": "Идентификатор счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                        }
                    },
                    {
                        "name": "modification",
                        "in": "path",
                        "description": "Изменение баланса",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "-500.00"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/TransactionCreate"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transaction/{id}/update": {
            "put": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Обновить транзакцию",
                "description": "Обновляет запись о транзакции и возвращает её",
                "operationId": "53af45dfdf96830a6076396d82af771a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Существующий идентификатор транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1bbf9e65-a2ec-46e8-b155-5dac60831817"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Новый идентификатор транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "bf03e5f6-b32c-4c3c-9435-6c4900c5ee22"
                        }
                    },
                    {
                        "name": "arendator_id",
                        "in": "query",
                        "description": "Новый идентификатор арендатора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                        }
                    },
                    {
                        "name": "bill_id",
                        "in": "query",
                        "description": "Новый идентификатор счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                        }
                    },
                    {
                        "name": "modification",
                        "in": "query",
                        "description": "Новое изменение счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "-500.00"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transaction/{id}/delete": {
            "delete": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Удалить транзакцию",
                "description": "Удаляет транзакцию",
                "operationId": "78f1e8935c2328e773e5c3107eb59106",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "deb4ff7a-c16b-4b9f-98db-d3c4e3cda010"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "title": "Transaction",
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "c62bf9d4-865f-49da-ba56-ab4fa7528698"
                    },
                    "arendator_id": {
                        "type": "uuid",
                        "example": "ae05a42b-9464-43df-83ed-dea6acf36f41"
                    },
                    "bill_id": {
                        "type": "uuid",
                        "example": "4d380caf-18a7-31e8-93ed-e5d2b5d93251"
                    },
                    "modification": {
                        "type": "decimal",
                        "example": 200.22
                    }
                },
                "type": "object"
            },
            "TransactionCreate": {
                "title": "TransactionCreate",
                "properties": {
                    "arendator_id": {
                        "type": "uuid",
                        "example": "ae05a42b-9464-43df-83ed-dea6acf36f41"
                    },
                    "bill_id": {
                        "type": "uuid",
                        "example": "4d380caf-18a7-31e8-93ed-e5d2b5d93251"
                    },
                    "modification": {
                        "type": "decimal",
                        "example": 200.22
                    }
                },
                "type": "object"
            },
            "Response401": {
                "title": "Response401",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "Response403": {
                "title": "Response403",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                },
                "type": "object"
            },
            "Response404": {
                "title": "Response404",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}